added back internal flag
