+++ src/chains/index.ts
@@ -1,16 +1,15 @@
 import { Chain } from 'viem/chains';
 import { queryClient } from '@/react-query';
 import { backendNetworksQueryKey, BackendNetworksResponse } from '@/resources/metadata/backendNetworks';
-
-import buildTimeNetworks from '@/references/networks.json';
-
-const backendNetworks = queryClient.getQueryData<BackendNetworksResponse>(backendNetworksQueryKey()) ?? buildTimeNetworks;
-
 import { ChainId, BackendNetwork, BackendNetworkServices, chainHardhat, chainHardhatOptimism } from './types';
 import { transformBackendNetworksToChains } from './utils/backendNetworks';
 import { gasUtils } from '@/utils';
 import { useConnectedToHardhatStore } from '@/state/connectedToHardhat';
 import { IS_TEST } from '@/env';
+import buildTimeNetworks from '@/references/networks.json';
+
+// NOTE: Prefer runtime data from backendNetworksQueryKey, but fallback to buildTimeNetworks if needed
+const backendNetworks = queryClient.getQueryData<BackendNetworksResponse>(backendNetworksQueryKey()) ?? buildTimeNetworks;
 
 const BACKEND_CHAINS = transformBackendNetworksToChains(backendNetworks.networks);
